+++ packages/nuxt/package.json
@@ -31,6 +31,7 @@
   },
   "dependencies": {
     "@mockline/utils": "*",
+    "@mockline/types": "*",
     "@iconify-json/heroicons": "^1.2.1",
     "@iconify-json/lucide": "^1.2.12",
     "@nuxt/content": "^2.13.4",
