+++ packages/nuxt/src/runtime/components/App.vue
