+++ packages/nuxt/src/runtime/components/Card.vue
