+++ packages/nuxt/src/runtime/composables/defineShortcuts.ts
