+++ packages/nuxt/src/runtime/types/mockline.d.ts
@@ -1,15 +1,6 @@
-import type { ToasterProps } from 'vue-sonner'
+import type { MocklineConfig } from '@mockline/types'
 
 const colors = JSON.stringify(options.colors) as const
-
-type MocklineConfig = {
-  primary?: typeof colors[number]
-  canvas?: typeof colors[number]
-  toast?: {
-    position?: ToasterProps['position']
-    duration?: number
-  }
-}
 
 declare module 'nuxt/schema' {
   type AppConfigInput = {
