+++ packages/nuxt/src/runtime/utils/colors.ts
@@ -68,41 +68,3 @@ export const colors = [
 ]
 
 export type Color = typeof colors
-
-// export const availableColors = colors.filter(color => color !== 'canvas' && color !== 'primary')
-
-/**
- * Generate a safelist of color classes with variants for TailwindCSS.
- *
- * @returns The safelist configuration.
- */
-export function generateColorSafelist(): { pattern: RegExp, variants: string[] }[] {
-  const variants = ['hover', 'focus', 'active', 'group-hover']
-  const classes = ['bg', 'text', 'border', 'ring']
-
-  return ['primary', 'canvas', ...colors].flatMap(color =>
-    classes.map(type => ({
-      pattern: new RegExp(`^${ type }-${ color }-(a)?(1[0-2]|[1-9])$`),
-      variants,
-    }))
-  )
-}
-
-/**
- * Generate a color scale based on the given color.
- *
- * @param color - The color for which to generate the scale.
- * @returns The color scale configuration.
- */
-export function generateScale(color: Color): Record<string, string> {
-  const scale = Array.from({ length: 12 }, (_, i) => {
-    const id = i + 1
-    return [
-      [id, `var(--${color}-${id})`],
-      [`a${id}`, `var(--${color}-a${id})`],
-      ['DEFAULT', `var(--${color}-${color === 'canvas' ? '1' : '9'})`],
-    ]
-  }).flat()
-
-  return Object.fromEntries(scale)
-}
