+++ packages/types/README.md
@@ -1,0 +1,1 @@
+# Mockline Types
