+++ packages/types/eslint.config.js
@@ -1,0 +1,3 @@
+import { createConfig } from "@hrcd/eslint-config"
+
+export default createConfig({})
