--- packages/utils/eslint.config.js
@@ -1,8 +1,0 @@
-import { createConfig } from "@hrcd/eslint-config"
-
-export default createConfig({
-  typescript: {
-    strict: true,
-    any: true
-  }
-})
