--- packages/utils/package.json
@@ -1,78 +1,0 @@
-{
-  "name": "@mockline/utils",
-  "version": "1.5.2",
-  "license": "Apache-2.0",
-  "type": "module",
-  "publishConfig": {
-    "access": "public"
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mockline/mockline.git",
-    "directory": "packages/utils"
-  },
-  "bugs": {
-    "url": "https://github.com/mockline/mockline/issues"
-  },
-  "exports": {
-    ".": {
-      "types": "./dist/module.d.ts",
-      "import": "./dist/module.mjs",
-      "require": "./dist/module.cjs",
-      "style": "./dist/runtime/index.css"
-    },
-    "./runtime/*": "./dist/runtime/*"
-  },
-  "style": "./dist/runtime/index.css",
-  "main": "./dist/module.cjs",
-  "types": "./dist/types.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "automd": "automd",
-    "prepack": "nuxt-module-build build",
-    "dev": "nuxi dev playground",
-    "build": "nuxt-module-build build",
-    "dev:prepare": "nuxt-module-build build --stub && nuxt-module-build prepare && nuxi prepare playground",
-    "test": "bun lint && bun typecheck && bun prepack",
-    "lint": "eslint .",
-    "lint:fix": "eslint --fix .",
-    "typecheck": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@iconify-json/heroicons": "^1.2.1",
-    "@iconify-json/lucide": "^1.2.12",
-    "@nuxt/fonts": "^0.10.2",
-    "@nuxt/icon": "^1.6.1",
-    "@nuxt/kit": "3.14.159",
-    "@nuxt/schema": "3.14.159",
-    "@nuxtjs/color-mode": "^3.5.2",
-    "@tailwindcss/postcss": "4.0.0-alpha.31",
-    "@tailwindcss/typography": "^0.5.15",
-    "@tailwindcss/vite": "4.0.0-alpha.31",
-    "@vueuse/core": "^11.2.0",
-    "defu": "^6.1.4",
-    "pathe": "^1.1.2",
-    "tailwindcss": "4.0.0-alpha.31",
-    "vue-sonner": "^1.2.5"
-  },
-  "devDependencies": {
-    "@nuxt/module-builder": "0.8.4",
-    "automd": "^0.3.12",
-    "eslint": "^9.14.0",
-    "nuxt": "3.14.159",
-    "typescript": "5.6.3"
-  },
-  "description": "Mockline Utils is the minimal utility library for Mockline. To kickstart your project with some useful modules like color mode, tailwindcss, nuxt icons, and more.",
-  "keywords": [
-    "nuxt",
-    "nuxt3",
-    "nuxt-module",
-    "mockline",
-    "vue",
-    "vue3",
-    "tailwindcss",
-    "tailwind"
-  ]
-}
