--- packages/utils/playground/assets/style/main.css
@@ -1,9 +1,0 @@
-@import "@mockline/utils";
-
-@theme {
-  --font-family-newsreader: "Newsreader", serif;
-}
-
-.content {
-  @apply prose prose-sm prose-headings:font-bold prose-a:underline;
-}
