--- packages/utils/playground/components/Test.vue
@@ -1,13 +1,0 @@
-<script setup lang="ts">
-
-</script>
-
-<template>
-  <div class="test underline">
-    I'm a test
-  </div>
-</template>
-
-<style scoped>
-
-</style>
