--- packages/utils/playground/nuxt.config.ts
@@ -1,14 +1,0 @@
-export default defineNuxtConfig({
-  modules: ['../src/module'],
-  devtools: { enabled: true },
-  compatibilityDate: '2024-07-09',
-  css: ['~/assets/style/main.css'],
-  icon: {
-    customCollections: [
-      {
-        prefix: 'custom',
-        dir: './assets/icons'
-      },
-    ],
-  },
-})
