--- packages/utils/playground/tsconfig.json
@@ -1,3 +1,0 @@
-{
-  "extends": "./.nuxt/tsconfig.json"
-}
