--- packages/utils/renovate.json
@@ -1,5 +1,0 @@
-{
-  "extends": [
-    "github>hugorcd/renovate-config"
-  ]
-}
