--- packages/utils/src/runtime/components/App.vue
@@ -1,15 +1,0 @@
-<template>
-  <div>
-    <slot />
-  </div>
-</template>
-
-<style>
-html,
-body,
-#__nuxt,
-#__layout {
-  width: 100%;
-  height: 100%;
-}
-</style>
