--- packages/utils/src/runtime/index.css
@@ -1,3 +1,0 @@
-@variant dark (&:where(.dark, .dark *));
-
-@plugin "@tailwindcss/typography";
