--- packages/utils/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "./.nuxt/tsconfig.json",
-  "exclude": ["dist", "playground", "node_modules"]
-}
